+++ epam-promo/components/overlays/NotificationCard.tsx
@@ -1,18 +1,22 @@
-import { withMods } from '@epam/uui-core';
-import { NotificationCard as uuiNotificationCard, NotificationCardProps as uuiNotificationCardProps } from '@epam/uui';
+import { createSkinComponent, devLogger } from '@epam/uui-core';
+import * as uui from '@epam/uui';
 import { EpamPrimaryColor } from '../types';
 
 export interface NotificationCardMods {
+    /** NotificationCard color */
     color?: EpamPrimaryColor | 'gray60';
 }
 
-export type NotificationCardProps = Omit<uuiNotificationCardProps, 'color'> & NotificationCardMods;
+export type NotificationCardProps = uui.NotificationCardCoreProps & NotificationCardMods;
 
-export const NotificationCard = withMods<Omit<uuiNotificationCardProps, 'color'>, NotificationCardMods>(
-    uuiNotificationCard,
-    () => [],
-    (props) => ({
-        ...props,
-        color: props.color ?? null,
-    }),
+export const NotificationCard = createSkinComponent<uui.NotificationCardProps, NotificationCardProps>(
+    uui.NotificationCard,
+    (props) => {
+        if (__DEV__) {
+            if (props.color === 'gray60') {
+                devLogger.warn('(NotificationCard) The gray60 value of color is deprecated and will be removed in future release. Use color blue instead or consult with your design team.');
+            }
+        }
+        return { ...props, color: props.color };
+    },
 );